Feat: reorganize project for pip
