+++ fasttrader/signals.py
