+++ fasttrader/simulator.py
