+++ tests/app.py
@@ -6,9 +6,9 @@ import pandas as pd
 import pandas_ta as pta
 import matplotlib.pyplot as plt
 
-from core.data import Data
-from core.signals import SignalGenerator
-from core.simulator import Simulator
+from fasttrader.data import Data
+from fasttrader.signals import SignalGenerator
+from fasttrader.simulator import Simulator
 
 # DATA
 # create an instance of the Data class
