+++ tests/random_signals_driver.py
@@ -2,7 +2,7 @@ import os
 import sys
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 
-from core.signals import SignalGenerator
+from fasttrader.signals import SignalGenerator
 
 signal_generator = SignalGenerator()
